refactor: extract event listener types and implementations
Extracts the event-listener types, implementations and tests into
their own module.

Also moves `ObjectRef` tests to their module.

 Conflicts:
	atspi-common/src/events/mod.rs
